Split tests to multiple files
